requirements.txt
faster-whisper
SpeechRecognition
pyaudio
numpy
python-Levenshtein
pyttsx3
gTTS
playsound==1.2.2
mysql-connector-python
python-statemachine
sounddevice
nvidia-cudnn-cu12==9.5.0.50
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cublas-cu12==12.4.5.8
torch
silero-vad
transformers
librosa
